requirements.txt
vectorbt>=0.25.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
futu-api>=6.8.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
